requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
email-validator==2.2.0
celery==5.4.0
redis==5.1.1
pybreaker==1.2.0
grpcio==1.68.0
grpcio-tools==1.68.0
liboqs-python==0.12.0
psycopg2-binary==2.9.9
pika==1.3.2
aiohttp==3.10.10
beautifulsoup4==4.12.3
playwright==1.49.1
